Added header and file for Player
