+++ Main.cpp
@@ -48,50 +48,6 @@ public:
     void attack() { cout << name << " attacks with a powerful move!\n"; }
 };
 
-// Player class definition
-class Player
-{
-public:
-    // Attributes
-    string name;
-    Pokemon chosenPokemon;
-
-    // Default constructor
-    Player()
-    {
-        name = "Trainer";
-        chosenPokemon = Pokemon(); // Using the default Pokemon constructor
-    }
-
-    // Parameterized constructor
-    Player(string p_name, Pokemon p_chosenPokemon)
-    {
-        name = p_name;
-        chosenPokemon = p_chosenPokemon;
-    }
-
-    void ChoosePokemon(int choice)
-    {
-        switch ((PokemonChoice)choice)
-        {
-        case PokemonChoice::Charmander:
-            chosenPokemon = Pokemon("Charmander", PokemonType::Fire, 100);
-            break;
-        case PokemonChoice::Bulbasaur:
-            chosenPokemon = Pokemon("Bulbasaur", PokemonType::Grass, 100);
-            break;
-        case PokemonChoice::Squirtle:
-            chosenPokemon = Pokemon("Squirtle", PokemonType::Water, 100);
-            break;
-        default:
-            chosenPokemon = Pokemon("Pikachu", PokemonType::Electric, 100);
-            break;
-        }
-        cout << "Player " << name << " chose " << chosenPokemon.name << "!\n";
-        Utility::WaitForEnter(); // Wait for user to press Enter before proceeding
-    }
-};
-
 // ProfessorOak class definition
 class ProfessorOak
 {
